Link navbar search component with search page functionality
